nvmef: fix sector size long -> int, read latency benchmark

+++ src/main/java/com/ibm/disni/nvmef/spdk/NativeDispatcher.java
@@ -47,7 +47,7 @@ public class NativeDispatcher {
 
 	public native long _nvme_ns_get_size(long namespace);
 
-	public native long _nvme_ns_get_sector_size(long namespace);
+	public native int _nvme_ns_get_sector_size(long namespace);
 
 	public native int _nvme_ns_io_cmd(long namespace, long queuePair, long address, long linearBlockAddress,
 										int count, long completionAddress, boolean write);
